// src/interfaces/nfl/nflWeeklyStatsKick.d.ts
import { RawStatData } from './nflStats';

export interface RawWeeklyStatKickData extends RawStatData, KickData {}

export interface KickData {
    player_weekly_id: number,
    fg_made: number,
    fg_missed: number,
    fg_blocked: number,
    fg_long: number,
    fg_att: number,
    fg_pct: number | null,
    pat_made: number,
    pat_missed: number,
    pat_blocked: number,
    pat_att: number,
    pat_pct: number | null,
    fg_made_distance: number,
    fg_missed_distance: number,
    fg_blocked_distance: number,
    gwfg_att: number,
    gwfg_distance: number,
    gwfg_made: number,
    gwfg_missed: number,
    gwfg_blocked: number,
    fg_made_0_19: number,
    fg_made_20_29: number,
    fg_made_30_39: number,
    fg_made_40_49: number,
    fg_made_50_59: number,
    fg_made_60_: number,
    fg_missed_0_19: number,
    fg_missed_20_29: number,
    fg_missed_30_39: number,
    fg_missed_40_49: number,
    fg_missed_50_59: number,
    fg_missed_60_: number,
    fg_made_list: string,
    fg_missed_list: string,
    fg_blocked_list: string,
}